requirements.txt
python-dotenv
ufiles
shazamio